requirements.txt
Django==2.1.7
djangorestframework==3.9.2
numpy==1.16.2
pytz==2018.9
scikit-learn==0.20.3
scipy==1.2.1
sklearn==0.0
pyowm==2.10.0
